Put in a custom acceleration
Remaps the arbitrary values returned by the 'Scrolling Speed' slider in the system prefs, to a value between 1 and 8 (for each notch in the slider).
This value is then used as a multiplier for the accel variable, which is increased at a rate of 0.25 per event. Unless the time between events is more than 400 ms, at which point it resets the accel to zero.
This is completely and totally a personal preference, and way better than the wacky shit in MacOS. I like it as it still allows a fast scroll if you are fast enough to continue the action.